Refactor Class App
